[QOL-8053] refactor
- Extract helper functions to a separate module to avoid circular dependency during initialisation
