+++ ckanext/datarequests/common.py
@@ -1,0 +1,43 @@
+# -*- coding: utf-8 -*-
+
+# Copyright (c) 2021 Queensland Government
+
+# This file is part of CKAN Data Requests Extension.
+
+# CKAN Data Requests Extension is free software: you can redistribute it and/or
+# modify it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+
+# CKAN Data Requests Extension is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+
+# You should have received a copy of the GNU Affero General Public License
+# along with CKAN Data Requests Extension. If not, see <http://www.gnu.org/licenses/>.
+
+from ckan.common import config
+import ckan.lib.helpers as h
+
+
+def get_config_bool_value(config_name, default_value=False):
+    value = config.get(config_name, default_value)
+    value = value if type(value) == bool else value != 'False'
+    return value
+
+
+def is_fontawesome_4():
+    if hasattr(h, 'ckan_version'):
+        ckan_version = float(h.ckan_version()[0:3])
+        return ckan_version >= 2.7
+    else:
+        return False
+
+
+def get_plus_icon():
+    return 'plus-square' if is_fontawesome_4() else 'plus-sign-alt'
+
+
+def get_question_icon():
+    return 'question-circle' if is_fontawesome_4() else 'question-sign'
